SIT-1408: template-wday2snow-worker-bidirectional-sync (Migration to 3.7).
